half-working
technically it works now, just an error that pops up because the card was deleted. i plan to switch it with an animation, the card doesn't get deleted, and the round is just so that the player wins. because this is a pretty weak mechanic in of itself, i'll try to focus a lot on the visuals and audio design to compensate, maybe